Refactor collectd binary to separate Unix-specific code.
Move OS-specific code from collectd.c into cmd.c.
This makes it easier to use the non-Linux-specific code when we port to Windows.
